Refactor : PointedText 컴포넌트 분리 및 경로 수정
