+++ pages/techblog/components/techDetailCard.tsx
@@ -8,7 +8,7 @@ import { formatDate } from '@utils/formatDate';
 
 import { useDropdownStore } from '@stores/dropdownStore';
 
-import SearchInput from '@components/common/searchInput';
+import SearchInput from '@components/common/techSearchInput/searchInput';
 
 import TechHeaderImg from '@public/image/techblog/TechHeaderImg.png';
