+++ pages/techblog/index.page.tsx
@@ -13,7 +13,7 @@ import { useObserver } from '@hooks/useObserver';
 
 import { Dropdown } from '@components/common/dropdowns/dropdown';
 import MobileDropdown from '@components/common/dropdowns/mobileDropdown';
-import SearchInput from '@components/common/searchInput';
+import SearchInput from '@components/common/techSearchInput/searchInput';
 import {
   MobileTechSkeletonList,
   TechSkeletonList,
